Removed some warning and problems for gcc
git-svn-id: file:///Volumes/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@14827 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/set/channel/bool.hpp
@@ -102,7 +102,7 @@ namespace Gecode { namespace Set { namespace Channel {
         Gecode::Int::BoolView::schedule(home, *this, Gecode::Int::ME_BOOL_VAL);
       View::schedule(home, *this, y.assigned() ? ME_SET_VAL : ME_SET_BB);
       if (y.assigned()) {
-        if (y.glbSize() == y.glbMax()-y.glbMin()+1) {
+        if (y.glbSize()==static_cast<unsigned int>(y.glbMax()-y.glbMin()+1)) {
           new (&delta) SetDelta(y.glbMin(),y.glbMax(),1,0);
         } else {
           new (&delta) SetDelta(2,0,1,0);
